+++ CHANGES.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 - Dependencies: Unpin commons-codec, to always use the latest version
+- Dependencies: Unpin lorrystream, to always use the latest version
 
 ## 2024/08/19 v0.0.17
 - Processor: Updated Kinesis Lambda processor to understand AWS DMS