+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 
 ## Unreleased
+
+## 2024/08/21 v0.0.18
 - Dependencies: Unpin commons-codec, to always use the latest version
 - Dependencies: Unpin lorrystream, to always use the latest version
 - MongoDB: Improve type mapper by discriminating between